validate/requirements.txt
Django==2.2.4
mysqlclient==1.4.4
numpy==1.17.4
opencv-python==4.1.2.30
Pillow>=8.0
pytz==2019.2
sqlparse==0.3.0
tk-tools
